SEAT_Project/apiGateway/requirements.txt
flask
grpcio
grpcio-tools
geopy
haversine
certifi